try simpler code refactor
