--- packages/kit-headless/src/components/carousel/carousel-context-id.ts
@@ -1,6 +1,0 @@
-import { createContextId } from '@builder.io/qwik';
-import { type CarouselContext } from './carousel-context.type';
-
-const CarouselContextId = createContextId<CarouselContext>('carousel-context');
-
-export default CarouselContextId;
