+++ packages/kit-headless/src/components/carousel/index.ts
@@ -1,7 +1,7 @@
-export { HCarousel as Root } from './carousel';
-export { HCarouselView as View } from './carousel-viewport';
-export { HCarouselContainer as Container } from './carousel-container';
-export { HCarouselSlide as Slide } from './slide';
-export { HCarouselPrev as Prev } from './prev-button';
-export { HCarouselNext as Next } from './next-button';
-export { HCarouselPagination as Pagination } from './pagination';
+export { CarouselRoot as Root } from './root';
+export { CarouselView as View } from './viewport';
+export { CarouselContainer as Container } from './container';
+export { CarouselSlide as Slide } from './slide';
+export { CarouselPrevious as Previous } from './previous';
+export { CarouselNext as Next } from './next';
+export { CarouselPagination as Pagination } from './pagination';
