--- packages/kit-headless/src/components/carousel/types.ts
@@ -1,3 +1,0 @@
-import { type PropsOf } from '@builder.io/qwik';
-
-export type CarouselButtonProps = PropsOf<'button'>;
